+++ backend/src/main/java/it/at/restfs/storage/Container.java
@@ -13,7 +13,7 @@ public class Container {
     private UUID id;
     
     //HttpStatusCode => NumberOf
-    private Map<Integer, Integer> statistics = new HashMap<>();
+    private Map<Integer, Long> statistics = new HashMap<>();
     
     private boolean statsEnabled = false;
